requirements.txt
aws-cdk-lib==2.21.1